backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
redis==5.0.1
celery==5.3.4
channels==4.0.0
channels-redis==4.1.0
drf-spectacular==0.26.5
Pillow==11.0.0
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
django-rest-knox==4.2.0
